tx history scope
