+++ lib/coinbase/client/api/external_addresses_api.rb
@@ -94,85 +94,6 @@ module Coinbase::Client
       return data, status_code, headers
     end
 
-    # List transactions for an address.
-    # List all transactions that interact with the address.
-    # @param network_id [String] The ID of the blockchain network
-    # @param address_id [String] The ID of the address to fetch the transactions for.
-    # @param [Hash] opts the optional parameters
-    # @option opts [Integer] :limit A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 10.
-    # @option opts [String] :page A cursor for pagination across multiple pages of results. Don&#39;t include this parameter on the first call. Use the next_page value returned in a previous response to request subsequent results.
-    # @return [AddressTransactionList]
-    def list_address_transactions(network_id, address_id, opts = {})
-      data, _status_code, _headers = list_address_transactions_with_http_info(network_id, address_id, opts)
-      data
-    end
-
-    # List transactions for an address.
-    # List all transactions that interact with the address.
-    # @param network_id [String] The ID of the blockchain network
-    # @param address_id [String] The ID of the address to fetch the transactions for.
-    # @param [Hash] opts the optional parameters
-    # @option opts [Integer] :limit A limit on the number of objects to be returned. Limit can range between 1 and 100, and the default is 10.
-    # @option opts [String] :page A cursor for pagination across multiple pages of results. Don&#39;t include this parameter on the first call. Use the next_page value returned in a previous response to request subsequent results.
-    # @return [Array<(AddressTransactionList, Integer, Hash)>] AddressTransactionList data, response status code and response headers
-    def list_address_transactions_with_http_info(network_id, address_id, opts = {})
-      if @api_client.config.debugging
-        @api_client.config.logger.debug 'Calling API: ExternalAddressesApi.list_address_transactions ...'
-      end
-      # verify the required parameter 'network_id' is set
-      if @api_client.config.client_side_validation && network_id.nil?
-        fail ArgumentError, "Missing the required parameter 'network_id' when calling ExternalAddressesApi.list_address_transactions"
-      end
-      # verify the required parameter 'address_id' is set
-      if @api_client.config.client_side_validation && address_id.nil?
-        fail ArgumentError, "Missing the required parameter 'address_id' when calling ExternalAddressesApi.list_address_transactions"
-      end
-      if @api_client.config.client_side_validation && !opts[:'page'].nil? && opts[:'page'].to_s.length > 5000
-        fail ArgumentError, 'invalid value for "opts[:"page"]" when calling ExternalAddressesApi.list_address_transactions, the character length must be smaller than or equal to 5000.'
-      end
-
-      # resource path
-      local_var_path = '/v1/networks/{network_id}/addresses/{address_id}/transactions'.sub('{' + 'network_id' + '}', CGI.escape(network_id.to_s)).sub('{' + 'address_id' + '}', CGI.escape(address_id.to_s))
-
-      # query parameters
-      query_params = opts[:query_params] || {}
-      query_params[:'limit'] = opts[:'limit'] if !opts[:'limit'].nil?
-      query_params[:'page'] = opts[:'page'] if !opts[:'page'].nil?
-
-      # header parameters
-      header_params = opts[:header_params] || {}
-      # HTTP header 'Accept' (if needed)
-      header_params['Accept'] = @api_client.select_header_accept(['application/json']) unless header_params['Accept']
-
-      # form parameters
-      form_params = opts[:form_params] || {}
-
-      # http body (model)
-      post_body = opts[:debug_body]
-
-      # return_type
-      return_type = opts[:debug_return_type] || 'AddressTransactionList'
-
-      # auth_names
-      auth_names = opts[:debug_auth_names] || []
-
-      new_options = opts.merge(
-        :operation => :"ExternalAddressesApi.list_address_transactions",
-        :header_params => header_params,
-        :query_params => query_params,
-        :form_params => form_params,
-        :body => post_body,
-        :auth_names => auth_names,
-        :return_type => return_type
-      )
-
-      data, status_code, headers = @api_client.call_api(:GET, local_var_path, new_options)
-      if @api_client.config.debugging
-        @api_client.config.logger.debug "API called: ExternalAddressesApi#list_address_transactions\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
-      end
-      return data, status_code, headers
-    end
-
     # Get the balances of an external address
     # List all of the balances of an external address
     # @param network_id [String] The ID of the blockchain network
