--- lib/coinbase/client/models/feature.rb
@@ -1,43 +1,0 @@
-=begin
-#Coinbase Platform API
-
-#This is the OpenAPI 3.0 specification for the Coinbase Platform APIs, used in conjunction with the Coinbase Platform SDKs.
-
-The version of the OpenAPI document: 0.0.1-alpha
-Contact: yuga.cohler@coinbase.com
-Generated by: https://openapi-generator.tech
-Generator version: 7.7.0
-
-=end
-
-require 'date'
-require 'time'
-
-module Coinbase::Client
-  class Feature
-    TRANSFER = "transfer".freeze
-    TRADE = "trade".freeze
-    FAUCET = "faucet".freeze
-    SERVER_SIGNER = "server_signer".freeze
-    UNKNOWN_DEFAULT_OPEN_API = "unknown_default_open_api".freeze
-
-    def self.all_vars
-      @all_vars ||= [TRANSFER, TRADE, FAUCET, SERVER_SIGNER, UNKNOWN_DEFAULT_OPEN_API].freeze
-    end
-
-    # Builds the enum from string
-    # @param [String] The enum value in the form of the string
-    # @return [String] The enum value
-    def self.build_from_hash(value)
-      new.build_from_hash(value)
-    end
-
-    # Builds the enum from string
-    # @param [String] The enum value in the form of the string
-    # @return [String] The enum value
-    def build_from_hash(value)
-      return value if Feature.all_vars.include?(value)
-      raise "Invalid ENUM value #{value} for class #Feature"
-    end
-  end
-end
